requirements_enhanced.txt
# Enhanced requirements for neuromorphic robot SNN project
# Install with: pip install -r requirements_enhanced.txt

# Core dependencies (same as original)
numpy>=1.21.0
torch>=2.0.0
snntorch>=0.6.0
matplotlib>=3.5.0
scipy>=1.9.0
pyyaml>=6.0
tqdm>=4.64.0

# Optional: experiment tracking
wandb>=0.15.0  # Weights & Biases (optional, for experiment tracking)

# Optional: advanced visualization (for better animations/gifs)
pillow>=9.0.0
imageio>=2.22.0
imageio-ffmpeg>=0.4.0  # for video export

# Optional: performance profiling
tensorboard>=2.12.0
psutil>=5.9.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
